refactor(select): centralize shared select contracts
