+++ src/renderer/src/types/select.ts
@@ -1,0 +1,10 @@
+import type {
+  SelectSuggestion,
+  SelectOption as UiSelectOption,
+} from "@aymurai/ui";
+
+export type SelectOption = UiSelectOption & {
+  shortText?: string;
+};
+
+export type { SelectSuggestion };
